requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==3.2
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
gunicorn==23.0.0
iniconfig==2.0.0
mongomock==4.1.2
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pymongo==4.8.0
pytest==8.3.2
pytest-django==4.8.0
python-decouple==3.8
pytz==2024.1
sentinels==1.0.0
sqlparse==0.2.4
typing_extensions==4.12.2
whitenoise==6.7.0
